+++ src/components/list-goods/List-goods.tsx
@@ -1,46 +1,21 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import Goods from '../goods/Goods';
 import styled from 'styled-components';
-import { loadProductsBest, loadProductsRecent } from '../../api/product';
 import { Link } from 'react-router-dom';
+import { GoodsProps } from '../../types/goods';
 
-const ListGoods = (text: any) => {
-  const [products, setProducts] = useState([]);
+interface IProps {
+  data: GoodsProps[];
+  title: string;
+}
 
-  const loadRecentData = async () => {
-    try {
-      const response = await loadProductsRecent();
-      const temp: any = response.data;
-      setProducts(temp.data);
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  const loadBestData = async () => {
-    try {
-      const response = await loadProductsBest();
-      const temp: any = response.data;
-      setProducts(temp.data);
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  useEffect(() => {
-    if (text.title === 'New ITEMS') {
-      loadRecentData();
-    } else if (text.title === 'Best ITEMS') {
-      loadBestData();
-    }
-  }, []);
-
+const ListGoods = ({ data, title }: IProps) => {
   return (
     <Section>
       <Wrapper>
         <ListTop>
           <TitleWapper>
-            <span>{text.title}</span>
+            <span>{title}</span>
           </TitleWapper>
           <More>
             <button>더보기</button>
@@ -48,7 +23,7 @@ const ListGoods = (text: any) => {
         </ListTop>
         <Divide></Divide>
         <ListBottom>
-          {products.map((item, idx) => (
+          {data.map((item, idx) => (
             <Link to={`products/${item.id}`}>
               <Goods data={item} key={idx}></Goods>
             </Link>
